의존 객체를 별도의 파일로 분리한다
