+++ lib/mk_resume/font_manager.rb
@@ -1,0 +1,20 @@
+module MkResume
+  class FontManager
+    FONT_SIZE = {:name => 11, :channel => 10, :heading => 9.5, :body => 9.5}
+
+    def load_font(pdf_doc)
+      Prawn::Font::AFM.hide_m17n_warning = true
+      pdf_doc.font_families.update(
+        "NotoSans" => {
+          normal: "./fonts/NotoSansKR-Regular.ttf",
+          bold: "./fonts/NotoSansKR-Bold.ttf"
+        }
+      )
+      pdf_doc.font "NotoSans"
+    end
+
+    def find_font_size usage
+      FONT_SIZE[usage]
+    end
+  end
+end
